[update] 重构honey-oss主服务 上传接口 [bugFix] 修复 dubbo 无法整通的bug

+++ honey-oss-server/src/main/java/com/eboy/honey/oss/server/application/po/FilePo.java
@@ -34,7 +34,7 @@ public class FilePo {
     /**
      * 文件大小
      */
-    private int fileSize;
+    private long fileSize;
     /**
      * 文件相对路径
      */
@@ -42,7 +42,7 @@ public class FilePo {
     /**
      * 分片大小
      */
-    private int shardSize;
+    private long shardSize;
     /**
      * 分片总数
      */
